+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.15.0-rc.11",
+  sdkVersion: "4.15.0-rc.12",
   genVersion: "2.933.0",
-  userAgent: "speakeasy-sdk/typescript 4.15.0-rc.11 2.933.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.15.0-rc.12 2.933.0 2.0 @orq-ai/node",
 } as const;