+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-01T11:45:49.928Z",
+    "2026-09-02T06:41:47.199Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {